// src/shared/api/generated/models/eventQueryPolicyDocumentDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Lola Backend Platform
 * CMS, integration, chat and realtime API for Lola AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { EventQueryPolicyItemDto } from "./eventQueryPolicyItemDto";

export interface EventQueryPolicyDocumentDto {
  enabled: boolean;
  /** @maxItems 50 */
  items: EventQueryPolicyItemDto[];
}